+++ packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @dotenv-cloak/core
 
+## 2.0.0
+
+### Major Changes
+
+- [`cd1481c`](https://github.com/yonycalsin/dotenv-cloak/commit/cd1481c5f3b451dddc0b0214562473e6bb826d1c) Thanks [@yonycalsin](https://github.com/yonycalsin)! - feat(core): parse and stringify the env vars
+
+### Patch Changes
+
+- Updated dependencies [[`376a945`](https://github.com/yonycalsin/dotenv-cloak/commit/376a945d3029cad508becb690ea9bf812f72740a), [`989561e`](https://github.com/yonycalsin/dotenv-cloak/commit/989561e0f785330e55ec9ae38553fd5498398d9d)]:
+  - @dotenv-cloak/stringify@3.0.0
+  - @dotenv-cloak/config@2.0.0
+
 ## 1.0.0
 
 ### Major Changes